feat: enhance book service with improved logging, error messages, and validation
- Introduced logging using SLF4J in BookService to log actions and exceptions.
- Added centralized error messages via an `ErrorMessage` enum for better consistency.
- Refactored validation methods to improve clarity and reusability.
- Switched `rating` field to `double` in `BookResponse`.
- Updated `BookController` to remove unnecessary optional handling.
- Configured centralized logging using `logback-spring.xml`.

+++ src/main/java/com/mello/nathalia/booksapi/api/response/BookResponse.java
@@ -5,5 +5,5 @@ public record BookResponse(
         String title,
         String author,
         String category,
-        int rating
+        double rating
 ){}
